Add Order Lifetimes
Ideally, we'd have an implementation where agents are able to cancel specific orders. But that requires the agents knowing what orders are open.

This brings us to the other thing that isn't implemented yet: reading market info. This is elaborated in comments in src/lib.rs.

Anyway, so we have order cancellation automatically which is sufficient for a first test run. I've stitched together a quick animation of the orderbook over time in matplotlib