Implement Agents
Currently using Bus crate to broadcast market info to agents.

Agents are parameterised by two weights. One determing how much to adjust reservation price based on quantity and the other determining much to scale spread based on variance of order book

There is issue with using Bus to implement broadcast because broadcast will only drop a message when all receivers have viewed it. And so since we are using try_broadcast, it means some new messages will be skipped when the box is full. Ideally, we would have a message is dropped when the buffer is full. and any reads take the latest message

Price data is still a bit wonky. I imagined that having agents submit orders based on market info would help smooth out the price action by providing some form of memory but that's not the case.